requirements.txt
openai
pillow
python-docx
pandas
fitz
PyMuPDF
chainlit
openpyxl
chromadb